Add CIDR support to TrustProxy and RobotIP directives
This works by adding new Vend::CIDR module to turn CIDR formats 11.22.33.44/5 into a regular
expression corresponding to the valid IPs in this range.  Due to the CIDRs being checked against
real, valid IPs, we are intentionally not making these patterns as complicated as they might be (for
instance, we are just checking for any wildcard digit to be /\d{1,3}/ rather than the much more
complicated regex `/2[0-4]\d|25[0-5]|1\d\d|\d|\d\d/` to explicitly validate the range of bytes
here.  Because we are comparing against structurally valid IPs, we consider the simplified form to
be much appreciated.

We also add new parse routine to detect CIDRs for the indicated directives and only apply this logic
in those cases.  Because we only perform special mapping on the individual fields which already
match the CIDR format, this should remain backward-compatible with any user-specified regexes in these
directives in the unfortunate case that they're using shell-style wildcards instead of regular
regexes.

Vend::CIDR code adapted from the original Javascript at: http://mog.xenowire.net/index.php?cidr2regex