Fix bug with mixed plaintext/crypted passwords on promotion
The password promotion code in UserDB did not work if `UserDB crypt 1` was set and passwords were
still set in plaintext; essentially there was no fallback unless the `from_plain` option was set,
which can only work if all the passwords are marked as plaintext (which if some passwords are hashed
would incorrectly/only expect the password value to be the actual hash), so this behavior was in
fact broken.

This patch makes the fallback for any non-recognized hash type be equivalent to 'from_plain', which
fixes the issues with the promotion.

Due to how this all works, there is still an unfortunate situation wherein if a user has a plaintext
password length that matches one of the known hash types, it will not work; in practice, this will
be the default method of length 13 (i.e., crypt).  This is a preexisting bug in the promote code and
due to the way this all works, we cannot distinguish between a 13-char password and a 13-char crypt
digest, so this can never be fixed in a way that will work in either scenario (short of checking
plaintext as a fallback to a failed crypt check).

Based on investigations by Mark Johnson and myself.