Fix response header parsing bug that failed to treat text output as UTF-8
It affected e.g. calls like this:

$::Tag->deliver('text/html', { status => '200' }, 'response body');

whereas this worked fine:

$::Tag->deliver('text/html', {}, 'response body');

not because of the status itself, but because that made the Content-Type
header no longer the beginning of $Vend::StatusLine, so it failed to
match the regex ^ which should be checking for the beginning of any
line, not the beginning of the string.