Make $Document robust when headers already exist
Usage patterns commonly seen in Interchange like this:

$Document->header("Content-type: text/plain");

don't work when headers are already there, and will break later header
additions, because there need to be newlines before and after the
new header.

Now we always add them, and the canon_status() call, which is always
used before sending headers in the response, cleans up the extra.