Allow MINIVEND_SOCKET to set the socket path at runtime.
Check the MINIVEND_SOCKET environment variable in vlink and vlink.pl and if set
use it to override the compiled-in path to the socket file.  This allows vlink
to be pre-compiled but the socket path can be changed at run time so that vlink
does not have to be re-compiled by the user.

Example usage in an apache config file:

SetEnv MINIVEND_SOCKET /path/to/new/socket
ScriptAlias / /path/to/vlink