Revert "fix an issue when session host = localhost"
This reverts commit f1b4335e3a2c3fb80be4fb87335b9b83e3d33f05.

The underlying issue was that the http server (HTTP::Server::PSGI) was
setting REMOTE_HOST to localhost and interchange is using that in the
session name without trying to resolve it to an ip address.

Starlet also sets REMOTE_HOST to localhost

Both HTTP::Server::PSGI and Starlet set REMOTE_ADDR to the internal
docker container ip address.

Starman sets both REMOTE_HOST and REMOTE_ADDR to the internal
docker container ip address which resolves the issue without
the need for this commit.