Fix database session locks
The LOCKLOOP block was missing a while statement which
according to Perl docs means that it is semantically identical
to a loop that executes once, and a next statement will
simply exit the block early.
So, when a session lock was found, it exited out of the LOCKLOOP
block, ending up in "# Should never get here", bypassing the lock.

Additionally, the $locktime and $pid variables were not being cleared
when the lock was cleared, causing the current pid to write a new
lock and use it to block itself indefinitely.